version 26.05.26.21.32 - phase8 extract service ping manager and discovery
