+++ MistHelper.py
@@ -17385,1061 +17385,57 @@ class GatewayHaExporter:
         print()  # Blank line after table for readability
 
 
+def _get_service_ping_manager_instance():  # type: ignore[no-untyped-def]
+    """Create extracted ServicePingManager instance with MistHelper runtime dependencies."""
+    from src.websocket.service_ping_manager import ServicePingManager as _SPM
+    from src.websocket.service_ping_manager import configure_service_ping_manager_dependencies as _configure_spm
+
+    _configure_spm(
+        apisession_dependency=apisession,
+        mistapi_dependency=mistapi,
+        prompt_utils=PromptUtils,
+        input_utils=InputUtils,
+        websocket_manager_class=WebSocketManager,
+        is_debug_mode=is_debug_mode,
+        api_tenant_fetch_utils=APITenantFetchUtils,
+        config_utils=ConfigUtils,
+        api_fetch_utils=APIFetchUtils,
+    )
+
+    return _SPM()
+
+
 class ServicePingManager:
-    """
-    Manager class for SSR Service Ping operations via WebSocket.
+    """Service ping manager (Menu 120).
 
-    Handles tenant/service discovery, user interaction, and WebSocket execution
-    for service-specific ping operations on SSR gateways.
+    Implementation extracted to `src/websocket/service_ping_manager.py` and
+    `src/websocket/service_ping_discovery.py`. This wrapper preserves
+    MistHelper menu orchestration and delegation only.
     """
 
-    # Default values for service ping operations
-    DEFAULT_HOST = "8.8.8.8"
-    DEFAULT_COUNT = 4
-    DEFAULT_SIZE = 56
-    MIN_SIZE = 56
-    MAX_SIZE = 65535
-    DEFAULT_TENANT = "testing-tools"
-    DEFAULT_SERVICE = "web-session"
+    from src.websocket.service_ping_manager import ServicePingManager as _Extracted
+
+    DEFAULT_HOST = _Extracted.DEFAULT_HOST
+    DEFAULT_COUNT = _Extracted.DEFAULT_COUNT
+    DEFAULT_SIZE = _Extracted.DEFAULT_SIZE
+    MIN_SIZE = _Extracted.MIN_SIZE
+    MAX_SIZE = _Extracted.MAX_SIZE
+    DEFAULT_TENANT = _Extracted.DEFAULT_TENANT
+    DEFAULT_SERVICE = _Extracted.DEFAULT_SERVICE
 
     def __init__(self):  # type: ignore[no-untyped-def]
-        """Initialize ServicePingManager with debug mode detection."""
-        self.debug_mode = is_debug_mode()  # type: ignore[no-untyped-call]
-        self.site_id: str | None = None
-        self.device_id: str | None = None
-        self.device_info: dict | None = None  # type: ignore[type-arg]
-        self.websocket_manager: WebSocketManager | None = None
-
-        # Tenant sources (in precedence order)
-        self.org_tenants: list[str] = []
-        self.site_tenants: list[str] = []
-        self.policy_tenants: list[str] = []
-        self.template_tenants: list[str] = []
-        self.device_tenants: list[str] = []
-
-        # Service sources
-        self.org_services: list[dict] = []  # type: ignore[type-arg]
-        self.org_service_names: list[str] = []
-        self.device_services: list[str] = []
-
-    def _debug_print(self, message: str) -> None:
-        """Print debug message if debug mode is enabled."""
-        if self.debug_mode:
-            print(f"[DEBUG] {message}")
-
-    def _select_site_and_device(self) -> bool:
-        """
-        Prompt user to select site and gateway device.
-
-        Returns:
-            bool: True if selection successful, False otherwise
-        """
-        self.site_id = PromptUtils.select_site_id_from_csv()
-        if not self.site_id:
-            print("! No site selected. Operation cancelled.")
-            return False
-
-        self._debug_print(f"Selected site_id = {self.site_id}")
-
-        self.device_id = PromptUtils.select_device_id_from_inventory(self.site_id, device_type="gateway")
-        if not self.device_id:
-            print("! No gateway devices found or selected. Service Ping requires an SSR gateway.")
-            return False
-
-        self._debug_print(f"Selected device_id = {self.device_id}")
-        return True
-
-    def _fetch_device_info(self) -> bool:
-        """
-        Fetch and validate device information.
-
-        Returns:
-            bool: True if device is valid for service ping, False otherwise
-        """
-        try:
-            logging.info("Fetching gateway device details for site %s and device %s", self.site_id, self.device_id)
-            rawdata = mistapi.api.v1.sites.devices.listSiteDevices(apisession, self.site_id, type="gateway").data
-            logging.debug("Retrieved %d gateway devices for site %s", len(rawdata or []), self.site_id)
-            self.device_info = next((device for device in rawdata if device.get("id") == self.device_id), None)
-            logging.debug("Device lookup complete for %s, found=%s", self.device_id, self.device_info is not None)
-        except Exception as error:
-            logging.warning(f"Could not retrieve device details: {error}")
-            self._debug_print(f"Device details error: {error}")
-            return self._confirm_unknown_device()
-
-        if not self.device_info:
-            return self._confirm_unknown_device()
-
-        return self._validate_device_type()
-
-    def _validate_device_type(self) -> bool:
-        """Validate device type and prompt for confirmation if not a gateway."""
-        if self.device_info is None:
-            return self._confirm_unknown_device()
-
-        device_type = self.device_info.get("type", "unknown")
-        device_model = self.device_info.get("model", "unknown")
-
-        self._debug_print(f"Device type: {device_type}, model: {device_model}")
-
-        if device_type == "gateway":
-            print(f"!? SSR Gateway detected (Model: {device_model})")
-            print("   -> Service Ping allows ping packets to follow service-specific paths")
-            return True
-
-        # Non-gateway device - warn and confirm
-        type_messages = {"ap": "Access Point", "switch": "Switch"}
-        device_label = type_messages.get(device_type, "Unknown device type")
-        print(f"!? WARNING: {device_label} detected (Model: {device_model})")
-        print("   -> Service Ping is designed for SSR gateways")
-
-        return self._confirm_proceed()
-
-    def _confirm_unknown_device(self) -> bool:
-        """Confirm proceeding with unknown device type."""
-        print("!? Cannot determine device type - proceeding with caution")
-        print("   -> Service Ping is designed for SSR gateways")
-        return self._confirm_proceed()
-
-    def _confirm_proceed(self) -> bool:
-        """Prompt user to confirm proceeding with non-optimal device."""
-        choice = (
-            InputUtils.safe_input(
-                "   -> Continue anyway? (y/N): ",
-                context="service_ping_continue",
-            )
-            .strip()
-            .lower()
-        )
-        if choice != "y":
-            print("Operation cancelled.")
-            return False
-        return True
-
-    def _fetch_all_tenants(self) -> None:
-        """Fetch tenants from all sources."""
-        print("\n-> Fetching organization tenants...")
-        self._fetch_org_tenants()
-
-        print("-> Fetching site tenants...")
-        self._fetch_site_tenants()
-
-        print("-> Fetching service policy tenants...")
-        self._fetch_policy_tenants()
-
-        print("-> Fetching gateway template tenants...")
-        self._fetch_template_tenants()
-
-    def _fetch_org_tenants(self) -> None:
-        """Fetch organization-level tenants."""
-        _tenant_utils = APITenantFetchUtils(
-            apisession, ConfigUtils.get_cached_or_prompted_org_id
-        )  # Create instance with runtime session and org ID resolver
-        tenants = _tenant_utils.organization_tenants()  # Fetch org-level tenants via injected session
-        if tenants:
-            self.org_tenants = tenants
-            print(f"   -> Found {len(self.org_tenants)} organization-level tenants")
-            self._debug_print(f"Organization tenants: {self.org_tenants}")
-        else:
-            print("   -> No organization-level tenants found")
-
-    def _fetch_site_tenants(self) -> None:
-        """Fetch site-level tenants."""
-        if self.site_id is None:
-            return
-        _tenant_utils = APITenantFetchUtils(
-            apisession, ConfigUtils.get_cached_or_prompted_org_id
-        )  # Create instance with runtime session and org ID resolver
-        tenants = _tenant_utils.site_tenants(self.site_id)  # Fetch site-level tenants via injected session
-        if tenants:
-            self.site_tenants = tenants
-            print(f"   -> Found {len(self.site_tenants)} site-level tenants")
-            self._debug_print(f"Site tenants: {self.site_tenants}")
-        else:
-            print("   -> No site-level tenants found")
-
-    def _fetch_policy_tenants(self) -> None:
-        """Fetch service policy tenants."""
-        _tenant_utils = APITenantFetchUtils(
-            apisession, ConfigUtils.get_cached_or_prompted_org_id
-        )  # Create instance with runtime session and org ID resolver
-        tenants = _tenant_utils.service_policy_tenants(
-            self.site_id
-        )  # Fetch service policy tenants via injected session
-        if tenants:
-            self.policy_tenants = tenants
-            print(f"   -> Found {len(self.policy_tenants)} service policy tenants")
-            self._debug_print(f"Service policy tenants: {self.policy_tenants}")
-        else:
-            print("   -> No service policy tenants found")
-
-    def _fetch_template_tenants(self) -> None:
-        """Fetch gateway template tenants."""
-        _tenant_utils = APITenantFetchUtils(
-            apisession, ConfigUtils.get_cached_or_prompted_org_id
-        )  # Create instance with runtime session and org ID resolver
-        tenants = _tenant_utils.gateway_template_tenants(
-            self.site_id
-        )  # Fetch gateway template tenants via injected session
-        if tenants:
-            self.template_tenants = tenants
-            print(f"   -> Found {len(self.template_tenants)} gateway template tenants")
-            self._debug_print(f"Gateway template tenants: {self.template_tenants}")
-        else:
-            print("   -> No gateway template tenants found")
-
-    def _fetch_all_services(self) -> None:
-        """Fetch services from organization and device configuration."""
-        print("\n-> Fetching organization services...")
-        self._fetch_org_services()
-
-        print("-> Fetching device configuration for additional options...")
-        self._fetch_device_config()
-
-    def _fetch_org_services(self) -> None:
-        """Fetch organization-level services."""
-        services = APIFetchUtils.organization_services()
-        if services:
-            self.org_services = services
-            self.org_service_names = [svc["name"] for svc in services if svc.get("name")]
-            print(f"   -> Found {len(self.org_service_names)} organization-level services")
-            self._debug_print(f"Organization services: {self.org_service_names}")
-        else:
-            print("   -> No organization-level services found")
-
-    def _fetch_device_config(self) -> None:
-        """Fetch device configuration for tenants and services."""
-        try:
-            logging.info("Fetching device configuration for site %s device %s", self.site_id, self.device_id)
-            config_response = mistapi.api.v1.sites.devices.getSiteDevice(apisession, self.site_id, self.device_id)
-            device_config = getattr(config_response, "data", {})
-            logging.debug("Device configuration retrieved with %d top-level keys", len(device_config.keys()))
-            self._debug_print(f"Device config keys: {list(device_config.keys())}")
-
-            self._extract_from_device_config(device_config)
-            self._extract_from_device_stats()
-
-        except Exception as error:
-            logging.warning(f"Could not fetch device configuration: {error}")
-            self._debug_print(f"Config error: {error}")
-            print("!? Cannot retrieve device configuration")
-
-    def _extract_from_device_config(self, config: dict) -> None:  # type: ignore[type-arg]
-        """Extract tenants and services from device configuration."""
-        tenants_set: set = set()  # type: ignore[type-arg]
-        services_set: set = set()  # type: ignore[type-arg]
-
-        # Extract from service_policies
-        for policy in config.get("service_policies", []):
-            if isinstance(policy, dict):
-                if tenant := policy.get("tenant"):
-                    tenants_set.add(tenant)
-                self._extract_services_from_policy(policy, services_set)
-
-        # Extract from routing_instances
-        for instance in config.get("routing_instances", []):
-            if isinstance(instance, dict):
-                if name := instance.get("name"):
-                    if not name.startswith("_"):
-                        tenants_set.add(name)
-
-        # Extract from router configuration
-        self._extract_from_router_config(config.get("router", {}), tenants_set, services_set)
-
-        # Filter and store results
-        self.device_tenants = sorted([t for t in tenants_set if t and not t.startswith("_")])
-        self.device_services = sorted([s for s in services_set if s and not s.startswith("_")])
-
-        self._report_device_config_results()
-
-    def _extract_services_from_policy(self, policy: dict, services_set: set) -> None:  # type: ignore[type-arg]
-        """Extract services from a service policy."""
-        services = policy.get("services", [])
-        if not isinstance(services, list):
-            return
-
-        for service_item in services:
-            if isinstance(service_item, dict):
-                if name := service_item.get("name"):
-                    services_set.add(name)
-            elif isinstance(service_item, str):
-                services_set.add(service_item)
-
-    def _extract_from_router_config(self, router: dict, tenants: set, services: set) -> None:  # type: ignore[type-arg]
-        """Extract tenants and services from router configuration."""
-        if not isinstance(router, dict):
-            return
-
-        for tenant_item in router.get("tenants", []):
-            if isinstance(tenant_item, dict):
-                if name := tenant_item.get("name"):
-                    tenants.add(name)
-
-        for service_item in router.get("services", []):
-            if isinstance(service_item, dict):
-                if name := service_item.get("name"):
-                    services.add(name)
-
-    def _extract_from_device_stats(self) -> None:
-        """Extract additional services from device stats."""
-        try:
-            logging.info(
-                "Fetching device stats for service discovery on site %s device %s",
-                self.site_id,
-                self.device_id,
-            )
-            stats_response = mistapi.api.v1.sites.stats.getSiteDeviceStats(apisession, self.site_id, self.device_id)
-            stats_data = getattr(stats_response, "data", {})
-            logging.debug("Device stats retrieved with %d top-level keys", len(stats_data.keys()))
-            self._debug_print(f"Stats keys: {list(stats_data.keys())}")
-
-            for service_stat in stats_data.get("service_stat", []):
-                if isinstance(service_stat, dict):
-                    if name := service_stat.get("name"):
-                        if not name.startswith("_") and name not in self.device_services:
-                            self.device_services.append(name)
-                            self.device_services.sort()
-
-        except Exception as error:
-            logging.debug(
-                "Device stats retrieval failed for site %s device %s: %s",
-                self.site_id,
-                self.device_id,
-                error,
-            )
-            self._debug_print(f"Could not fetch stats: {error}")
-
-    def _report_device_config_results(self) -> None:
-        """Report device configuration extraction results."""
-        self._debug_print(f"Found tenants from device: {self.device_tenants}")
-        self._debug_print(f"Found services from device: {self.device_services}")
-
-        if self.device_tenants:
-            print(f"   -> Found {len(self.device_tenants)} tenants from device configuration")
-        else:
-            print("   -> No tenants found in device configuration")
-
-        if self.device_services:
-            print(f"   -> Found {len(self.device_services)} additional services from device configuration")
-        else:
-            print("   -> No additional services found in device configuration")
-
-    def _build_combined_tenants(self) -> list[str]:
-        """Combine tenants from all sources with deduplication."""
-        combined = list(self.org_tenants)
-
-        # Add from each source if not already present
-        for source in [self.site_tenants, self.policy_tenants, self.template_tenants, self.device_tenants]:
-            for tenant in source:
-                if tenant not in combined:
-                    combined.append(tenant)
-
-        # Ensure default tenant exists
-        if self.DEFAULT_TENANT not in combined:
-            combined.append(self.DEFAULT_TENANT)
-            self._debug_print(f"Added default tenant: {self.DEFAULT_TENANT}")
-
-        self._debug_print(f"Combined tenant list: {combined}")
-        return combined
-
-    def _build_combined_services(self) -> list[str]:
-        """Combine services from all sources with deduplication."""
-        combined = list(self.org_service_names)
-
-        for service in self.device_services:
-            if service not in combined:
-                combined.append(service)
-
-        # Ensure default service exists
-        if self.DEFAULT_SERVICE not in combined:
-            combined.append(self.DEFAULT_SERVICE)
-            self._debug_print(f"Added default service: {self.DEFAULT_SERVICE}")
-
-        self._debug_print(f"Combined service list: {combined}")
-        return combined
-
-    def _prompt_for_tenant(self, available_tenants: list[str]) -> str | None:
-        """Prompt user to select a tenant."""
-        if not available_tenants:
-            return self._prompt_manual_tenant()
-
-        print("\nAvailable Tenants:")
-        self._display_tenant_categories(available_tenants)
-
-        default_index = None
-        if self.DEFAULT_TENANT in available_tenants:
-            default_index = available_tenants.index(self.DEFAULT_TENANT)
-
-        print(f"  [{len(available_tenants)}] Skip tenant selection")
-
-        return self._get_tenant_selection(available_tenants, default_index)
-
-    def _display_tenant_categories(self, all_tenants: list[str]) -> None:  # noqa: C901
-        """Display tenants organized by source category."""
-        index = 0
-
-        # Organization tenants
-        if self.org_tenants:
-            print(f"  Organization Tenants ({len(self.org_tenants)}):")
-            for name in self.org_tenants:
-                print(f"    [{index}] {name} (org networks)")
-                index += 1
-
-        # Site-only tenants
-        site_only = [t for t in self.site_tenants if t not in self.org_tenants]
-        if site_only:
-            print(f"  Site Tenants ({len(site_only)}):")
-            for name in site_only:
-                print(f"    [{index}] {name} (site networks)")
-                index += 1
-
-        # Policy-only tenants
-        policy_only = [t for t in self.policy_tenants if t not in self.org_tenants and t not in self.site_tenants]
-        if policy_only:
-            print(f"  Service Policy Tenants ({len(policy_only)}):")
-            for name in policy_only:
-                print(f"    [{index}] {name} (service policies)")
-                index += 1
-
-        # Template-only tenants
-        template_only = [
-            t
-            for t in self.template_tenants
-            if t not in self.org_tenants and t not in self.site_tenants and t not in self.policy_tenants
-        ]
-        if template_only:
-            print(f"  Gateway Template Tenants ({len(template_only)}):")
-            for name in template_only:
-                print(f"    [{index}] {name} (gateway templates)")
-                index += 1
-
-        # Device-only tenants
-        device_only = [
-            t
-            for t in self.device_tenants
-            if t not in self.org_tenants
-            and t not in self.site_tenants
-            and t not in self.policy_tenants
-            and t not in self.template_tenants
-        ]
-        if device_only:
-            print(f"  Device Configuration Tenants ({len(device_only)}):")
-            for name in device_only:
-                print(f"    [{index}] {name} (device config)")
-                index += 1
-
-        # Remaining (defaults)
-        remaining = [
-            t
-            for t in all_tenants
-            if t not in self.org_tenants
-            and t not in self.site_tenants
-            and t not in self.policy_tenants
-            and t not in self.template_tenants
-            and t not in self.device_tenants
-        ]
-        if remaining:
-            print(f"  Additional Tenants ({len(remaining)}):")
-            for name in remaining:
-                print(f"    [{index}] {name} (default/custom)")
-                index += 1
-
-    def _get_tenant_selection(self, tenants: list[str], default_index: int | None) -> str | None:
-        """Get tenant selection from user input."""
-        while True:
-            try:
-                prompt = f"\nSelect tenant index (0-{len(tenants)})"
-                if default_index is not None:
-                    prompt += f" [default: {default_index} ({self.DEFAULT_TENANT})]: "
-                else:
-                    prompt += " [default: skip]: "
-
-                selection = InputUtils.safe_input(prompt, context="service_ping_tenant_selection").strip()
-
-                if not selection:
-                    if default_index is not None:
-                        tenant = tenants[default_index]
-                        print(f"!? Using default tenant: {tenant}")
-                        return tenant
-                    return None
-
-                idx = int(selection)
-                if idx == len(tenants):
-                    print("!? Skipping tenant selection")
-                    return None
-
-                if 0 <= idx < len(tenants):
-                    tenant = tenants[idx]
-                    self._print_tenant_source(tenant)
-                    return tenant
-
-                print(f"Please enter a number between 0 and {len(tenants)}")
-
-            except ValueError:
-                print("Please enter a valid number")
-            except KeyboardInterrupt:
-                print("\nOperation cancelled")
-                raise
-
-    def _print_tenant_source(self, tenant: str) -> None:
-        """Print the source of a selected tenant."""
-        if tenant in self.org_tenants:
-            print(f"!? Selected organization tenant: {tenant}")
-        elif tenant in self.site_tenants:
-            print(f"!? Selected site tenant: {tenant}")
-        elif tenant in self.policy_tenants:
-            print(f"!? Selected service policy tenant: {tenant}")
-        elif tenant in self.template_tenants:
-            print(f"!? Selected gateway template tenant: {tenant}")
-        elif tenant in self.device_tenants:
-            print(f"!? Selected device configuration tenant: {tenant}")
-        else:
-            print(f"!? Selected default/custom tenant: {tenant}")
-
-    def _prompt_manual_tenant(self) -> str | None:
-        """Prompt for manual tenant entry when none available."""
-        print("\n-> No tenants found in any configuration source")
-        manual = InputUtils.safe_input(
-            "-> Enter tenant name manually (or press Enter to skip): ",
-            context="service_ping_tenant_manual",
-        ).strip()
-        if manual:
-            print(f"!? Manual tenant: {manual}")
-            return manual
-        print("-> Proceeding without tenant (may cause service ping to fail)")
-        return None
-
-    def _prompt_for_service(self, available_services: list[str]) -> str:
-        """Prompt user to select a service."""
-        if not available_services:
-            return self._prompt_required_service()
-
-        print("\nAvailable Services:")
-        self._display_service_categories(available_services)
-
-        default_index = None
-        if self.DEFAULT_SERVICE in available_services:
-            default_index = available_services.index(self.DEFAULT_SERVICE)
-
-        print(f"  [{len(available_services)}] Enter custom service name")
-
-        return self._get_service_selection(available_services, default_index)
-
-    def _display_service_categories(self, all_services: list[str]) -> None:
-        """Display services organized by source category."""
-        index = 0
-
-        # Organization services
-        if self.org_service_names:
-            print(f"  Organization Services ({len(self.org_service_names)}):")
-            for name in self.org_service_names:
-                details = next((s for s in self.org_services if s["name"] == name), {})
-                svc_type = details.get("type", "custom")
-                desc = details.get("description", "")
-                if desc:
-                    print(f"    [{index}] {name} ({svc_type}) - {desc}")
-                else:
-                    print(f"    [{index}] {name} ({svc_type})")
-                index += 1
-
-        # Device-only services
-        device_only = [s for s in self.device_services if s not in self.org_service_names]
-        if device_only:
-            print(f"  Device Configuration Services ({len(device_only)}):")
-            for name in device_only:
-                print(f"    [{index}] {name} (device config)")
-                index += 1
-
-        # Remaining (defaults)
-        remaining = [s for s in all_services if s not in self.org_service_names and s not in self.device_services]
-        if remaining:
-            print(f"  Additional Services ({len(remaining)}):")
-            for name in remaining:
-                print(f"    [{index}] {name} (default/custom)")
-                index += 1
-
-    def _get_service_selection(self, services: list[str], default_index: int | None) -> str:
-        """Get service selection from user input."""
-        while True:
-            try:
-                prompt = f"\nSelect service index (0-{len(services)}) or enter custom"
-                if default_index is not None:
-                    prompt += f" [default: {default_index} ({self.DEFAULT_SERVICE})]: "
-                else:
-                    prompt += ": "
-
-                selection = InputUtils.safe_input(prompt, context="service_ping_service_selection").strip()
-
-                if not selection:
-                    if default_index is not None:
-                        service = services[default_index]
-                        print(f"!? Using default service: {service}")
-                        return service
-                    print("Please enter a service name or select from the list")
-                    continue
-
-                try:
-                    idx = int(selection)
-                    if idx == len(services):
-                        return self._prompt_custom_service()
-
-                    if 0 <= idx < len(services):
-                        service = services[idx]
-                        self._print_service_source(service)
-                        return service
-
-                    print(f"Please enter a number between 0 and {len(services)}")
-
-                except ValueError:
-                    # Treat as custom service name
-                    print(f"!? Custom service: {selection}")
-                    return selection
-
-            except KeyboardInterrupt:
-                print("\nOperation cancelled")
-                raise
-
-    def _print_service_source(self, service: str) -> None:
-        """Print the source of a selected service with details if available."""
-        if service in self.org_service_names:
-            print(f"!? Selected organization service: {service}")
-            details = next((s for s in self.org_services if s["name"] == service), {})
-            if details.get("description"):
-                print(f"  Description: {details['description']}")
-            if details.get("type"):
-                print(f"  Type: {details['type']}")
-        elif service in self.device_services:
-            print(f"!? Selected device configuration service: {service}")
-        else:
-            print(f"!? Selected default/custom service: {service}")
-
-    def _prompt_custom_service(self) -> str:
-        """Prompt for custom service name."""
-        while True:
-            service = InputUtils.safe_input(
-                "Enter custom service name: ",
-                context="service_ping_custom_service",
-            ).strip()
-            if service:
-                print(f"!? Custom service: {service}")
-                return service
-            print("Service name cannot be empty")
-
-    def _prompt_required_service(self) -> str:
-        """Prompt for required service when none available."""
-        print("\n-> No services found in organization or device configuration")
-        while True:
-            service = InputUtils.safe_input(
-                "Enter service name: ",
-                context="service_ping_required_service",
-            ).strip()
-            if service:
-                print(f"!? Custom service: {service}")
-                return service
-            print("Service is required. Please enter a service name.")
-
-    def _prompt_for_ping_parameters(self) -> dict:  # type: ignore[type-arg]
-        """Prompt for ping parameters and build payload."""
-        # Host
-        host = InputUtils.safe_input(
-            "\nEnter target host/IP to ping [default: 8.8.8.8]: ",
-            context="service_ping_host",
-        ).strip()
-        if not host:
-            host = self.DEFAULT_HOST
-            print(f"!? Using default destination: {host}")
-
-        # Count
-        count = self._prompt_for_count()
-
-        # Size
-        size = self._prompt_for_size()
-
-        # HA Node
-        node = self._prompt_for_node()
-
-        return {"host": host, "count": count, "size": size, "node": node}
-
-    def _prompt_for_count(self) -> int:
-        """Prompt for ping count."""
-        count_input = InputUtils.safe_input(
-            "Enter ping count [default: 4]: ",
-            context="service_ping_count",
-        ).strip()
-        try:
-            count = int(count_input) if count_input else self.DEFAULT_COUNT
-            return max(1, count)
-        except ValueError:
-            return self.DEFAULT_COUNT
-
-    def _prompt_for_size(self) -> int:
-        """Prompt for packet size."""
-        size_input = InputUtils.safe_input(
-            "Enter packet size in bytes [default: 56]: ",
-            context="service_ping_size",
-        ).strip()
-        try:
-            size = int(size_input) if size_input else self.DEFAULT_SIZE
-            return max(self.MIN_SIZE, min(size, self.MAX_SIZE))
-        except ValueError:
-            return self.DEFAULT_SIZE
-
-    def _prompt_for_node(self) -> str | None:
-        """Prompt for HA node selection."""
-        node_input = (
-            InputUtils.safe_input(
-                "Enter HA node (node0/node1) [optional]: ",
-                context="service_ping_node",
-            )
-            .strip()
-            .lower()
-        )
-        return node_input if node_input in ["node0", "node1"] else None
-
-    def _build_payload(self, service: str, tenant: str | None, params: dict) -> dict:  # type: ignore[type-arg]
-        """Build service ping API payload."""
-        payload = {"host": params["host"], "service": service, "count": params["count"], "size": params["size"]}
-
-        if tenant:
-            payload["tenant"] = tenant
-        if params["node"]:
-            payload["node"] = params["node"]
-
-        return payload
-
-    def _display_configuration(self, payload: dict) -> None:  # type: ignore[type-arg]
-        """Display service ping configuration summary."""
-        print("\n" + "-" * 50)
-        print("Service Ping Configuration:")
-        print(f"  Host: {payload['host']}")
-        print(f"  Service: {payload['service']}")
-        print(f"  Count: {payload['count']}")
-        print(f"  Size: {payload['size']} bytes")
-        if payload.get("tenant"):
-            print(f"  Tenant: {payload['tenant']}")
-        if payload.get("node"):
-            print(f"  HA Node: {payload['node']}")
-        print("-" * 50)
-
-        self._debug_validate_service(payload["service"])
-
-    def _debug_validate_service(self, service: str) -> None:
-        """Log debug information about service validation."""
-        if not self.debug_mode:
-            return
-
-        known_services = ["web-session", "LANS", "RBO_SSH"]
-        if service in known_services:
-            print(f"[DEBUG] Using known valid service: {service}")
-        else:
-            print(f"[DEBUG] Using custom service: {service} (may not exist on device)")
-
-    def _setup_websocket(self) -> bool:
-        """Initialize and connect WebSocket."""
-        self.websocket_manager = WebSocketManager(apisession)
-
-        if not self.websocket_manager.connect():
-            print("! Failed to establish WebSocket connection")
-            return False
-
-        self._debug_print("WebSocket connection established")
-
-        command_channel = f"/sites/{self.site_id}/devices/{self.device_id}/cmd"
-        if not self.websocket_manager.subscribe_to_channel(command_channel):
-            print("! Failed to subscribe to device command channel")
-            return False
-
-        self._debug_print(f"Subscribed to channel: {command_channel}")
-        print("-> WebSocket connected and subscribed")
-
-        print("-> Waiting for subscription confirmation...")
-        if not self.websocket_manager.wait_for_subscription_confirmation(command_channel, timeout_seconds=15):  # type: ignore[no-untyped-call]
-            print("! Subscription confirmation not received within timeout")
-            print("! Proceeding anyway, but results may not be received")
-        else:
-            print("-> Subscription confirmed")
-
-        return True
-
-    def _execute_service_ping(self, payload: dict) -> str | None:  # type: ignore[type-arg]
-        """Execute service ping API call and return session ID."""
-        print("-> Issuing Service Ping command...")
-
-        self._debug_print(f"Service ping payload being sent: {payload}")
-        logging.info(f"Sending service ping via mistapi to device: {self.device_id}")
-        logging.info(f"Service ping payload: {payload}")
-
-        try:
-            response = mistapi.api.v1.sites.devices.servicePingFromSsr(
-                apisession, self.site_id, self.device_id, payload
-            )
-
-            logging.info(f"Service ping mistapi response status: {response.status_code}")
-            logging.info(f"Service ping mistapi response data: {response.data}")
-
-            self._debug_print(f"mistapi Response Status = {response.status_code}")
-            self._debug_print(f"mistapi Response Data = {response.data}")
-
-            if response.status_code != 200:
-                print(f"Failed to issue Service Ping command. Status {response.status_code}: {response.data}")
-                logging.error(f"Service ping failed - status {response.status_code}: {response.data}")
-                return None
-
-            session_id = response.data.get("session", "")
-            if session_id:
-                short_id = session_id[:8] + "..." if len(session_id) > 8 else session_id
-                print(f"-> Service Ping command issued (session: {short_id})")
-                self._debug_print(f"Full session ID: {session_id}")
-            else:
-                print("-> Service Ping command issued (no session ID returned)")
-
-            return session_id or None
-
-        except Exception as error:
-            print(f"Error issuing Service Ping command via mistapi: {error}")
-            logging.error(f"Service ping error: {error}")
-            self._debug_print(f"mistapi exception details: {type(error).__name__}: {error}")
-            return None
-
-    def _wait_for_results(self, session_id: str) -> dict | None:  # type: ignore[type-arg]
-        """Wait for WebSocket results."""
-        print("-> Waiting for Service Ping results...")
-
-        self._debug_print(f"Full session ID = {session_id}")
-        self._debug_print("Starting to wait for WebSocket results...")
-
-        if self.websocket_manager is None:
-            return None
-
-        # Determine timeout based on device type
-        if self.device_info and self.device_info.get("type") == "gateway":
-            timeout_seconds = 45
-            activity_timeout = 5
-            print("   -> Using extended timeout for SSR gateway (45s total, 5s activity)")
-        else:
-            timeout_seconds = 30
-            activity_timeout = 3
-
-        result = self.websocket_manager.wait_for_command_result(
-            session_id, timeout_seconds=timeout_seconds, activity_timeout_seconds=activity_timeout
-        )
-
-        self._debug_print(f"wait_for_command_result returned: {result is not None}")
-        if result:
-            self._debug_print(f"Result keys: {list(result.keys())}")
-
-        return result
-
-    def _display_results(self, result: dict | None, payload: dict) -> None:  # type: ignore[type-arg]
-        """Display service ping results."""
-        if result:
-            self._display_success_results(result, payload)
-        else:
-            self._display_timeout_results(payload)
-
-    def _display_success_results(self, result: dict, payload: dict) -> None:  # type: ignore[type-arg]
-        """Display successful ping results."""
-        print("\n" + "=" * 60)
-        print("SERVICE PING RESULTS:")
-        print("=" * 60)
-
-        if self.device_info is not None:
-            self._display_device_context(payload)
-
-        raw_output = result.get("raw", "")
-        if raw_output:
-            print("PING OUTPUT:")
-            print("-" * 40)
-            print(raw_output)
-
-        parsed_output = result.get("Output", "")
-        if parsed_output and parsed_output != raw_output:
-            print("\nPARSED OUTPUT:")
-            print("-" * 40)
-            print(parsed_output)
-
-        if not raw_output and not parsed_output:
-            print("No output data received")
-            if self.device_info and self.device_info.get("type") != "gateway":
-                self._display_non_gateway_troubleshooting()
-
-        print("=" * 60)
-        self._log_success(payload)
-
-    def _display_device_context(self, payload: dict) -> None:  # type: ignore[type-arg]
-        """Display device context in results."""
-        if self.device_info is None:
-            return
-        device_type = self.device_info.get("type", "unknown")
-        device_model = self.device_info.get("model", "unknown")
-        device_name = self.device_info.get("name", "Unknown Device")
-
-        print(f"Device: {device_name} ({device_type.upper()}: {device_model})")
-        print(f"Service: {payload['service']} -> Host: {payload['host']}")
-
-        if device_type == "gateway":
-            print("Note: Service-specific routing path used for ping packets")
-        else:
-            print("Note: Device may not fully support service ping functionality")
-
-        print("-" * 60)
-
-    def _display_non_gateway_troubleshooting(self) -> None:
-        """Display troubleshooting for non-gateway devices."""
-        print("\nTroubleshooting for non-gateway devices:")
-        print("-> Service Ping is designed specifically for SSR gateways")
-        print("-> Try using regular ping (Menu 87) instead")
-        print("-> Verify device supports service ping functionality")
-
-    def _log_success(self, payload: dict) -> None:  # type: ignore[type-arg]
-        """Log successful operation."""
-        if self.device_info:
-            name = self.device_info.get("name", "Unknown Device")
-            dtype = self.device_info.get("type", "unknown")
-            logging.info(
-                "Service ping completed for %s (%s) - Service: %s, Host: %s",
-                name,
-                dtype,
-                payload["service"],
-                payload["host"],
-            )
-        else:
-            logging.info(
-                "Service ping completed for device %s - Service: %s, Host: %s",
-                self.device_id,
-                payload["service"],
-                payload["host"],
-            )
-
-    def _display_timeout_results(self, payload: dict) -> None:  # type: ignore[type-arg]
-        """Display timeout/no results message with troubleshooting."""
-        print("\nNo Service Ping results received within timeout period.")
-
-        if not self.device_info:
-            logging.warning(f"Service ping timeout - no results received for device {self.device_id}")
-            return
-
-        device_type = self.device_info.get("type", "unknown")
-        device_name = self.device_info.get("name", "Unknown Device")
-        print(f"Device: {device_name} ({device_type})")
-
-        troubleshooting = {
-            "gateway": [
-                "Verify service name is valid for this SSR",
-                "Check if host is reachable through the specified service",
-                "Confirm SSR routing configuration for the service",
-                "Try with a different service name",
-            ],
-            "switch": [
-                "Switches typically do not support service ping",
-                "Try using regular ping (Menu 87) for basic connectivity",
-                "Service ping is an SSR-specific feature",
-            ],
-            "ap": [
-                "Access Points do not support service ping",
-                "Try using regular ping (Menu 87) for basic connectivity",
-                "Service ping is an SSR-specific feature",
-            ],
-        }
-
-        tips = troubleshooting.get(
-            device_type,
-            ["Service ping is designed for SSR gateways", "Try using regular ping (Menu 87) for basic connectivity"],
-        )
-
-        if device_type == "gateway":
-            print("\nTroubleshooting for SSR gateways:")
-        else:
-            print(f"\nNote: {tips[0]}")
-            tips = tips[1:]
-
-        for tip in tips:
-            print(f"-> {tip}")
-
-        logging.warning(f"Service ping timeout - no results received for device {self.device_id}")
-
-    def _cleanup(self) -> None:
-        """Clean up WebSocket connection."""
-        try:
-            if self.websocket_manager is not None:
-                self.websocket_manager.disconnect()
-                print("-> WebSocket connection closed")
-        except Exception as error:
-            logging.warning(f"WebSocket cleanup error: {error}")
+        """Initialize wrapper by mirroring extracted manager state for test compatibility."""
+        extracted = _get_service_ping_manager_instance()
+        self.__dict__.update(extracted.__dict__)
+        self._delegate = extracted
+
+    def __getattr__(self, name):  # type: ignore[no-untyped-def]
+        """Delegate unknown attributes and methods to extracted implementation."""
+        return getattr(self._delegate, name)
 
     def execute(self) -> None:
-        """
-        Main execution method for service ping operation.
-
-        Orchestrates the complete service ping workflow:
-        1. Site and device selection
-        2. Device validation
-        3. Tenant and service discovery
-        4. User configuration prompts
-        5. WebSocket setup and execution
-        6. Results display and cleanup
-        """
-        logging.debug("ENTER: ServicePingManager.execute")
-
-        if self.debug_mode:
-            print("[DEBUG] Starting Service Ping via WebSocket operation...")
-            print(f"[DEBUG] Command line args: {sys.argv}")
-
-        try:
-            # Phase 1: Selection and validation
-            if not self._select_site_and_device():
-                return
-
-            if not self._fetch_device_info():
-                return
-
-            # Phase 2: Discovery
-            self._fetch_all_tenants()
-            self._fetch_all_services()
-
-            # Phase 3: Build combined lists
-            all_tenants = self._build_combined_tenants()
-            all_services = self._build_combined_services()
-
-            # Phase 4: User configuration
-            print("\n" + "=" * 50)
-            print("SERVICE PING CONFIGURATION")
-            print("=" * 50)
-
-            tenant = self._prompt_for_tenant(all_tenants)
-            service = self._prompt_for_service(all_services)
-            params = self._prompt_for_ping_parameters()
-
-            # Phase 5: Build and display payload
-            payload = self._build_payload(service, tenant, params)
-            self._display_configuration(payload)
-
-            # Phase 6: Execute
-            print(f"\n-> Executing Service Ping on device {self.device_id}...")
-
-            if not self._setup_websocket():
-                return
-
-            session_id = self._execute_service_ping(payload)
-            if not session_id:
-                print("! No session ID received - cannot wait for results")
-                return
-
-            # Phase 7: Wait and display results
-            result = self._wait_for_results(session_id)
-            self._display_results(result, payload)
-
-        except KeyboardInterrupt:
-            print("\nOperation cancelled by user")
-            logging.info("Service ping operation cancelled by user")
-
-        except Exception as error:
-            print(f"Error during Service Ping operation: {error}")
-            logging.error(f"Service ping error: {error}")
-
-        finally:
-            self._cleanup()
-            logging.debug("EXIT: ServicePingManager.execute")
+        """Execute menu 120 via the extracted service ping manager."""
+        self._delegate.execute()
 
 
 # ============================================================================
