+++ src/websocket/__init__.py
@@ -4,10 +4,14 @@ from src.websocket.commands import WebSocketCommands
 from src.websocket.context import WebSocketCmdDeps
 from src.websocket.diag_commands import WebSocketNetworkDiagCommands
 from src.websocket.manager import WebSocketManager
+from src.websocket.service_ping_discovery import ServicePingDiscoveryMixin
+from src.websocket.service_ping_manager import ServicePingManager
 
 __all__ = [
     "WebSocketCmdDeps",
     "WebSocketCommands",
     "WebSocketManager",
     "WebSocketNetworkDiagCommands",
+    "ServicePingDiscoveryMixin",
+    "ServicePingManager",
 ]
